DoAdvCommand: retail's route-length guard is a SIGNED compare (77.10 -> 77.17)
why-branch named it: retail tests the vector's byte span with
`test ecx,0xfffffffc; jle`, which is VC6 folding `(bytes>>2) <= 0` on a
SIGNED count. Dinkumware's size() is unsigned, so the bare
`size() <= 0` our CL saw folded to `je` instead. static_cast<int>
restores the signed sense (the cast floor bans casts to ENUM types, not
this one). Branch-shape distance 13 -> 11.

+++ src/advmgr.cpp
@@ -459,7 +459,7 @@ NewmapCell* advManager::DoAdvCommand(type_point* trigger_point)
         currHero->field_11c = 0;
         advWindow->SetSleepImage(0);
 
-        if (gpSearchArray->result.size() <= 0)
+        if (static_cast<int>(gpSearchArray->result.size()) <= 0)
             break;
 
         int savedShowRoute = bShowRoute;
